+++ tests/test1/conftest.py
@@ -27,6 +27,7 @@ def pytest_configure():
             'crispy_bootstrap5',
             'ordered_model',
             'django_tables2',
+            "django_object_detail",
             "crud_views_bootstrap5.apps.CrudViewsBootstrap5Config",
             "crud_views.apps.CrudViewsConfig",
             'tests.test1.app',